move DNSServerIdentity to StandaloneDNSProxy
DNSServerIdentity is only used by SDP so it is moved under
tree.
